# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StockSense inventory management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /operations/{kind}:
    get:
      operationId: listOperations
      tags: [operations]
      summary: List receipts, deliveries, transfers, or adjustments
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OperationKindPath"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/StatusQuery"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
      responses:
        "200":
          description: Operations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOperations"
    post:
      operationId: createOperation
      tags: [operations]
      summary: Create an inventory operation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OperationKindPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationInput"
      responses:
        "201":
          description: Operation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
  /operations/{kind}/{id}:
    get:
      operationId: getOperation
      tags: [operations]
      summary: Get an inventory operation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OperationKindPath"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
    patch:
      operationId: updateOperation
      tags: [operations]
      summary: Edit a draft operation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OperationKindPath"
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationUpdate"
      responses:
        "200":
          description: Updated operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
  /operations/{kind}/{id}/validate:
    post:
      operationId: validateOperation
      tags: [operations]
      summary: Complete an operation and apply stock changes
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OperationKindPath"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Validated operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "400":
          $ref: "#/components/responses/Error"
  /operations/{kind}/{id}/cancel:
    post:
      operationId: cancelOperation
      tags: [operations]
      summary: Cancel an operation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OperationKindPath"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Canceled operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
    OperationKindPath:
      name: kind
      in: path
      required: true
      schema:
        type: string
        enum: [receipts, deliveries, transfers, adjustments]
    SearchQuery:
      name: search
      in: query
      schema: { type: string }
...
    StatusQuery:
      name: status
      in: query
      schema:
        type: string
        enum: [DRAFT, WAITING, READY, DONE, CANCELED]
...
    PageQuery:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeQuery:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    OperationLine:
      type: object
      required: [id, productId, productName, sku, uom, quantity]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        sku: { type: string }
        uom: { type: string }
        quantity: { type: number, minimum: 0.01 }
        physicalQuantity: { type: ["number", "null"] }
        difference: { type: ["number", "null"] }
    Operation:
      type: object
      required: [id, number, kind, status, createdAt, createdBy, lines]
      properties:
        id: { type: integer }
        number: { type: string }
        kind:
          type: string
          enum: [RECEIPT, DELIVERY, TRANSFER, ADJUSTMENT]
        status:
          type: string
          enum: [DRAFT, WAITING, READY, DONE, CANCELED]
        partner: { type: ["string", "null"] }
        warehouseId: { type: ["integer", "null"] }
        warehouseName: { type: ["string", "null"] }
        sourceLocationId: { type: ["integer", "null"] }
        sourceLocationName: { type: ["string", "null"] }
        destinationWarehouseId: { type: ["integer", "null"] }
        destinationWarehouseName: { type: ["string", "null"] }
        destinationLocationId: { type: ["integer", "null"] }
        destinationLocationName: { type: ["string", "null"] }
        reason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        createdBy: { type: string }
        lines:
          type: array
          items: { $ref: "#/components/schemas/OperationLine" }
    OperationInput:
      type: object
      required: [lines]
      properties:
        partner: { type: string }
        warehouseId: { type: integer }
        sourceLocationId: { type: integer }
        destinationWarehouseId: { type: integer }
        destinationLocationId: { type: integer }
        reason: { type: string }
        lines:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: integer }
              quantity: { type: number, minimum: 0.01 }
              physicalQuantity: { type: number, minimum: 0 }
    OperationUpdate:
      allOf:
        - $ref: "#/components/schemas/OperationInput"
    PaginatedOperations:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Operation" }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...