# lib/api-zod/src/generated/types/operationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * StockSense inventory management API
 * OpenAPI spec version: 0.1.0
 */
import type { OperationInputLinesItem } from './operationInputLinesItem';

export interface OperationInput {
  partner?: string;
  warehouseId?: number;
  sourceLocationId?: number;
  destinationWarehouseId?: number;
  destinationLocationId?: number;
  reason?: string;
  /** @minItems 1 */
  lines: OperationInputLinesItem[];
}